+++ app/src/main/java/Entities/Implementations/VolumetricRecipeItem.java
@@ -14,16 +14,16 @@ public class VolumetricRecipeItem extends AbstractRecipeItem {
      * @param amount     The quantity of this ingredient as float.
      * @param optional   Is the ingredient optional?
      */
-    public VolumetricRecipeItem(Ingredient ingredient, float amount, boolean optional) {
+    public VolumetricRecipeItem(Ingredient ingredient, double amount, boolean optional) {
         super(ingredient, amount, optional);
     }
 
-    public VolumetricRecipeItem(UUID id, Ingredient ingredient, float amount, boolean optional) {
+    public VolumetricRecipeItem(UUID id, Ingredient ingredient, double amount, boolean optional) {
         super(id, ingredient, amount, optional);
     }
 
     public String display(){
-        float tempQuantity = this.quantity();
+        double tempQuantity = this.quantity();
         String stringQuantity;
 
         // checks if the quantity ends in .0 or .00 (whole number)
@@ -32,7 +32,7 @@ public class VolumetricRecipeItem extends AbstractRecipeItem {
             stringQuantity = Integer.toString((int) tempQuantity);
         } else {
             // Convert to a string for returning
-            stringQuantity = Float.toString(tempQuantity);
+            stringQuantity = Double.toString(tempQuantity);
         }
         return stringQuantity + "ml of " + this.ingredient().name();
     }
